Added deduplication of wrapped fortran MPI calls
Now that we are calling the actual fortran PMPI bindings within the
fortran wrappers (as opposed to converting to C-flavored args and
calling the C PMPI bindings), there is the possibility of
double-counting wrapped fortran MPI calls. Namely, if a particular MPI
implementation calls the C MPI bindings (which are wrapped by IPM) from
within the fortran PMPI bindings, then we end up intercepting the call
twice. This has been observed in both MPICH and OpenMPI.

As a first-cut solution, I've now added a flag that indicates whether we
are within a fortran PMPI call (set / unset within the fortran wrapper),
which can in turn be used to selectively call into the IPM_
instrumentation routine within the C wrapper.