Initial try at split fotran libs and call dedup
This commit implements two experimental changes:

1) Splitting the fortran intercept routines out of the main libipm
library and placing them in libipmf

This includes all automatically generated intercept routines, as well as
those in mpi_fortran.c. Intercepts for MPI_Init and MPI_Finalize are
retained in mpi_init.c and mpi_finalize.c, respectively (since they are
actually still forwarded to the associated C MPI bindings, not fortran
PMPI). The idea is to make it so that fortran PMPI bindings are only
needed when building fortran code (i.e. -lipm is used for C / C++ while
-lipmf -lipm is used for fortran).

2) Deduplication of fortran MPI call counts

Now that we are calling the actual fortran PMPI bindings within the
fortran wrappers (as opposed to converting to C-flavored args and
calling the C PMPI bindings), there is the possibility of double-
counting wrapped fortran MPI calls. Namely, if a particular MPI
implementation calls the C MPI bindings (which are wrapped by IPM) from
within the fortran PMPI bindings, then we end up intercepting the call
twice. This has been observed in both MPICH and OpenMPI.

As a first-cut solution, I've now added a flag that indicates whether we
are within a fortran PMPI call (set / unset within the fortran wrapper),
which can in turn be used to selectively call into the IPM_
instrumentation routine within the C wrapper.